Desktop/NT208.K21.ANTT.7-BE-develop/requirements.txt
django-extensions
PyJWT
djangorestframework
django-cors-headers
gunicorn
python-dotenv
psycopg2==2.7.5
django-heroku
whitenoise
Pillow